+++ rio/src/screen/messenger.rs
@@ -17,7 +17,7 @@ impl Messenger {
     }
 
     #[inline]
-    fn send_write<B: Into<Cow<'static, [u8]>>>(&self, data: B) {
+    pub fn send_write<B: Into<Cow<'static, [u8]>>>(&self, data: B) {
         let bytes = data.into();
         // terminal hangs if we send 0 bytes through.
         if bytes.len() == 0 {